see the commit of player.hpp and orc.hpp

+++ src/Global_entity_values.hpp
@@ -14,10 +14,10 @@ struct player_marker{}; // marker for the player
 struct name {std::string name;};
 struct pos {int pos_x; int pos_y; int pos_level;};
 struct icon {char icon;};
-struct fg {uint8_t fg_r; int fg_g; int fg_b;};
+struct fg {uint8_t fg_r; uint8_t fg_g; uint8_t fg_b;};
 struct hp {int hp;};
 struct atk {int atk;};
 struct xp {int xp; int level;};
-struct AI {int AI;};
+struct AI {uint8_t AI;};
 struct actions {struct game_actions actions;};
 #endif
